requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0

# TensorRT (ensure compatibility with your CUDA version)
#For cuda 13
tensorrt>=10.14.1.48
# Install separately: pip install nvidia-tensorrt

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.0

# Tracking
filterpy>=1.4.5
scikit-learn>=1.3.0
scipy>=1.11.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Monitoring & Logging
tensorboard>=2.15.0
wandb>=0.16.0
pynvml>=11.5.0
prometheus-client>=0.19.0

# Performance & Utilities
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Image processing
Pillow>=10.0.0

# YAML support
pyyaml>=6.0
